pipeline/requirements.txt
rasterio>=1.3
fiona>=1.9
shapely>=2.0
geopandas>=0.14
numpy>=1.24
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
tqdm>=4.64